purchase orders make it to warehousing context

+++ Warehousing/src/main/java/be/com/pashioya/mineralflowsystem/warehousing/ports/in/CreateInventoryItemCommand.java
@@ -5,5 +5,5 @@ package be.com.pashioya.mineralflowsystem.warehousing.ports.in;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
-public record CreateInventoryItemCommand(UUID customerUUID, UUID materialUUID, UUID warehouseUUID, double quantity, LocalDateTime dateReceived) {
+public record CreateInventoryItemCommand(UUID customerUUID, UUID materialUUID, UUID warehouseUUID, int quantity, LocalDateTime dateReceived) {
 }
